Add files via upload
The flood water level monitoring stake is a smart device designed to monitor and report flood water levels in real-time. This device can be installed along rivers, canals, or areas with high flood risk. Equipped with accurate water level sensors, it continuously provides information on the current water level, helping to predict floods and issue early warnings.

The monitoring stake is equipped with wireless connectivity, allowing data to be sent directly to a control center or mobile application. It also features solar-powered operation, ensuring sustainable performance even in harsh conditions. This system is an ideal solution for disaster prevention and response, protecting people and property from the impacts of floods